Fix bug on profile edit page
Appending edit/ to the URL of anyone's profile page would let you
edit their profile.

This was fixed by changing the rendered model on an edit page to
the model of the requesting user rather than the model of the
profile page.

The side effect is that the URL will be the viewed profile's edit
page instead of the user sending the request, even if the content
is.